exynos7885: Fix issues with the DPP_G0 HWC layer
There are flags for dealing with a problematic quirk of the DPP_G0 HWC
layer on 7885 that went unnoticed. These are defined in the exynos850
configuration, with values that are correct as-is for the 7885:

 #define DEDICATED_CHANNEL_TYPE MPP_DPP_G
 #define DEDICATED_CHANNEL_INDEX 0
 #define USE_DEDICATED_TOP_WINDOW

So, add the definitions, and allow the DPP_G0 layer to be usable. While
at it, also align the 7885 configuration with the 850 one, since it just
seems nicer.

There now appear to be no HWC-related issues on 7885.

Change-Id: I46f65f2eaa992e24ab334f63dd67fbc2e9a3510f